ADD: 캐러셀컴포넌트 렌더링
