+++ src/pages/Home/Home.jsx
@@ -3,6 +3,7 @@ import Header from './components/Header';
 import HomeEvent from './components/HomeEvent';
 import HomeGrid from './components/HomeGrid';
 import HomeMain from './components/HomeMain';
+import Carousel from '../../components/Carousel';
 
 const Home = () => {
   return (
@@ -10,6 +11,7 @@ const Home = () => {
       {/* 헤더 - src컴포넌트 */}
       <Header />
       <HomeMain />
+      <Carousel />
       <HomeGrid />
       <HomeEvent />
       {/* 푸터 - src컴포넌트*/}
